Firebase push notifications

+++ app/GraphQL/Mutations/DriverResolver.php
@@ -2,10 +2,12 @@
 
 namespace App\GraphQL\Mutations;
 
-use \App\Driver;
+use App\Driver;
 use App\DriverVehicle;
-use \App\Traits\UploadOneFile;
-use \App\Traits\DeleteOneFile;
+use App\DeviceToken;
+use App\Traits\UploadOneFile;
+use App\Traits\DeleteOneFile;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Support\Arr;
 use GraphQL\Type\Definition\ResolveInfo;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
@@ -29,7 +31,7 @@ class DriverResolver
         $input = collect($args)->except(['directive', 'avatar'])->toArray();
         $input['password'] = Hash::make($input['phone']);
 
-        if ($args['avatar']) {
+        if (array_key_exists('avatar', $args) && $args['avatar']) {
             $url = $this->uploadOneFile($args['avatar'], 'avatars');
             $input['avatar'] = $url;
         }
@@ -49,7 +51,7 @@ class DriverResolver
             throw new \Exception('The provided driver ID is not found.');
         }
 
-        if ($args['avatar']) {
+        if (array_key_exists('avatar', $args) && $args['avatar']) {
             if ($driver->avatar) $this->deleteOneFile($driver->avatar, 'avatars');
             $url = $this->uploadOneFile($args['avatar'], 'avatars');
             $input['avatar'] = $url;
@@ -79,6 +81,17 @@ class DriverResolver
         }
 
         $driver = auth('driver')->user();
+
+        if (array_key_exists('device_id', $args) && array_key_exists('platform', $args)) {
+            try {
+                DeviceToken::where('device_id', $args['device_id'])->firstOrFail();
+            } catch (ModelNotFoundException $e) {
+                $tokenInput = collect($args)->only(['platform', 'device_id'])->toArray();
+                $tokenInput['tokenable_id'] = $driver->id;
+                $tokenInput['tokenable_type'] = 'App\Driver';
+                DeviceToken::create($tokenInput);
+            }
+        }
 
         return [
             'access_token' => $token,
